src/core/connection-manager.ts
import type TransportManager from "../models/transport-manager.model";
import { Session, SessionState } from "../models/session.model";
import { Credentials, AuthResult } from "../models/auth.model";
import { AuthService } from "./auth-service";
import { v4 as uuidv4 } from "uuid";

/**
 * Manages database client connections and transport layer operations
 */
class ConnectionManager implements TransportManager {
  private sessions: Map<string, Session>;
  private maxSessions: number;
  private authService: AuthService;

  constructor(maxSessions: number = 100) {
    this.sessions = new Map();
    this.maxSessions = maxSessions;
    this.authService = new AuthService();
  }

  handleConnection(): Session {
    if (this.sessions.size >= this.maxSessions) {
      throw new Error("Maximum session limit reached");
    }

    const session: Session = {
      id: uuidv4(),
      createdAt: new Date(),
      state: SessionState.CONNECTING,
      lastActivity: new Date(),
    };

    this.sessions.set(session.id, session);
    return session;
  }

  async authenticateClient(credentials: Credentials): Promise<boolean> {
    const session = this.getSessionByCredentials(credentials);
    if (!session) {
      return false;
    }

    const authResult = await this.authService.authenticate(credentials);
    if (authResult.success && authResult.securityContext) {
      session.state = SessionState.ACTIVE;
      session.lastActivity = new Date();
      session.securityContext = authResult.securityContext;
      return true;
    }

    session.state = SessionState.TERMINATED;
    return false;
  }

  manageSession(): void {
    // Implementation will be added
  }

  loadBalance(): void {
    // Implementation will be added
  }

  // Helper methods
  getSessions(): Map<string, Session> {
    return this.sessions;
  }

  getSessionByCredentials(credentials: Credentials): Session | undefined {
    for (const session of this.sessions.values()) {
      if (
        session.credentials &&
        session.credentials.username === credentials.username
      ) {
        return session;
      }
    }
    return undefined;
  }
}

export default ConnectionManager;
